Rename path for `Upload` to the new path
The uploads are logged in an `uploads` table.
The path is the relative path the uploader returns. So in case of the
old paths this is `uploads/{user|project|...}` for the newly created
uploads this needs to be `uploads/system/{user|project|...}`

The namespace paths used by the fileuploader are
`uploads/{relative}/{namespace}/{path}`.