Make --prune a configurable parameter in fetching a git remote
By default, --prune is added to the command-line of a `git fetch`
operation, but for repositories with many references this can take
hours to run.

For Geo, we shouldn't see references on the secondary that aren't
on the primary, and including this option isn't worth the performance
penalty.

Relates to gitlab-org/gitlab-ee#5047