Default to putting a project in a group in tests
There are two kinds of namespaces used in GitLab: user and group namespaces.
The former is implied with a `type` of `nil`, and the latter uses a `type` of `Group`.
If a project namespace is created with the same ID as the user's namespace, there
will be ambiguity that can cause test failures.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/45350