EE : implement `rbac` for UpgradeCommand which is now required by BaseCommand.
This enables us to mount the right serviceAccountName in the pod when
we run the UpgradeComand in a RBAC-enabled cluster. The operation we are
attempting is to get a config map which will require the right priv in an RBAC cluster.

For now, there is no change to the helm command in the rbac case.
Setting `rbac.create` is not relevant here as that creates rbac roles
and rolebindings which should have already been performed on helm
install previously.