Update Clusters::Applications::Prometheus#prometheus_client to use the KubeClient wrapper correctly
It was by chance that prometheus_client called #proxy_url which then
proxied to the _first_ client of the KubeClient wrapper which happened
to be the core client (the one that has `/api` as the endpoint). Make it
expliclitly connect to the core client.

Improve the specs for #prometheus_client to stub the discovery HTTP
response and not the internals of the method under test (!)

Also, restrict the proxying (via method_missing) to only the methods
that start with Kubeclient::Client ENTITY_METHODS (get, create, etc)