Elasticsearch v6 support
In order to support Elasticsearch 6 and upwards, we need to migrate away
from using parent/child relationships and instead use the new JOIN
datatype. In order to do this we also need to switch to having a single
mapping type, as ES6 does not support more than one type per index.

Changes:

- changed all our document types to be `doc`
- implemented a custom type field in our elasticsearch documents,
called `type`, in order to be able to query by type.
- Elasticsearch's `_id` field now uses the model's `#es_id`, which is
the class name, an underscore, and the row ID.
- `elasticsearch-model` gem gets monkey-patched in order to use `#es_id`
to retrieve models from elasticsearch results