Redirect user back to secondary after logout/login
When you initially log in to a Geo secondary node, it redirects you
to the primary node, you log in there, and it then takes you back to
the secondary.

When you log out, it also takes you to the primary login screen.
However logging in again will log the user into the primary.

The changes introduced here redirect the user back to the secondary
node after logging into the primary.