Add reusable project_selector component
This commit adds a resuable UI component that allows a user to search
for a project name, shows the search results, and allows the user to
select one or more projects.  This component communicates with its
parent using props and events.

This component was originally created for use in the EE-specific
"Operations Dashboard" page, but it is applicable for CE use cases as
well, and so was added as a CE shared component.

In addition, some logic was extracted from the frequent_items_list_item
component into a shared mixin to avoid logic duplication.